% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkreadr-package.R
\docType{package}
\name{bulkreadr-package}
\alias{bulkreadr}
\alias{bulkreadr-package}
\title{bulkreadr: The Ultimate Tool for Reading Data in Bulk}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Designed to simplify and streamline the process of reading and processing large volumes of data in R, this package offers a collection of functions tailored for bulk data operations. It enables users to efficiently read multiple sheets from Microsoft Excel and Google Sheets workbooks, as well as various CSV files from a directory. The data is returned as organized data frames, facilitating further analysis and manipulation. Ideal for handling extensive data sets or batch processing tasks, bulkreadr empowers users to manage data in bulk effortlessly, saving time and effort in data preparation workflows. Additionally, the package seamlessly works with labelled data from SPSS and Stata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gbganalyst/bulkreadr}
  \item \url{https://gbganalyst.github.io/bulkreadr/}
  \item Report bugs at \url{https://github.com/gbganalyst/bulkreadr/issues}
}

}
\author{
\strong{Maintainer}: Ezekiel Ogundepo \email{gbganalyst@gmail.com} (\href{https://orcid.org/0000-0003-3974-2733}{ORCID})

Other contributors:
\itemize{
  \item Ernest Fokoue \email{epfeqa@rit.edu} (\href{https://orcid.org/0000-0002-0748-9166}{ORCID}) [contributor]
  \item Golibe Ezeechesi \email{golibe.ezeechesi@gmail.com} [contributor]
  \item Fatimo Adebanjo \email{adebanjofatimo2000@gmail.com} [contributor]
  \item Isaac Ajao \email{isaacoluwaseyiajao@gmail.com} [contributor]
}

}
\keyword{internal}
