% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_files_from_dir.R
\name{read_excel_files_from_dir}
\alias{read_excel_files_from_dir}
\title{Read Excel Workbooks Data from a Directory}
\usage{
read_excel_files_from_dir(dir_path, col_types = NULL, .id = NULL)
}
\arguments{
\item{dir_path}{Path to the directory containing the \code{xls}/\code{xlsx} files.}

\item{col_types}{Either \code{NULL} to guess all from the spreadsheet or a
character vector containing one entry per column from these options:
"skip", "guess", "logical", "numeric", "date", "text" or "list". If exactly
one \code{col_type} is specified, it will be recycled. The content of a cell in
a skipped column is never read and that column will not appear in the data
frame output. A list cell loads a column as a list of length 1 vectors,
which are typed using the type guessing logic from \code{col_types = NULL}, but
on a cell-by-cell basis.}

\item{.id}{The name of an optional identifier column. Provide a string to
create an output column that identifies each input. The column will use
names if available, otherwise it will use positions.}
}
\value{
A \link[tibble:tibble-package]{tibble}. If there is any column type mismatch during data frames row binding, an error will occur. This is because R cannot combine columns of different types. For example, you cannot combine a column of integers with a column of characters.
}
\description{
\code{read_excel_files_from_dir()} reads all Excel workbooks in the \code{"~/data"} directory and returns an appended dataframe.
}
\examples{
\donttest{
directory <- system.file("xlsxfolder",  package = "bulkreadr")

read_excel_files_from_dir(dir_path = directory, .id = "cut")
}
# Column types mismatch error --------------------------------------
# If the `read_excel_files_from_dir()` function complains about a data type mismatch,
# then set the `col_types` argument to `"text"`.
# This will make all the column types in the resulting dataframe be characters.


}
\seealso{
\code{\link[=read_excel_workbook]{read_excel_workbook()}} which imports data from multiple sheets of an Excel workbook
}
