% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkreadr-package.R
\docType{package}
\name{bulkreadr-package}
\alias{bulkreadr}
\alias{bulkreadr-package}
\title{bulkreadr: The Ultimate Tool for Reading Data in Bulk}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Designed to simplify and streamline the process of reading and processing large volumes of data in R. With a collection of functions tailored for bulk data operations, the package allows users to efficiently read multiple sheets from 'Microsoft Excel'/'Google Sheets' workbooks and multiple CSV files from a directory. It returns the data as organized data frames, making it convenient for further analysis and manipulation. Whether dealing with extensive data sets or batch processing tasks, 'bulkreadr' empowers users to effortlessly handle data in bulk, saving time and effort in data preparation workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gbganalyst/bulkreadr}
  \item Report bugs at \url{https://github.com/gbganalyst/bulkreadr/issues}
}

}
\author{
\strong{Maintainer}: Ezekiel Ogundepo \email{gbganalyst@gmail.com} (\href{https://orcid.org/0000-0003-3974-2733}{ORCID})

Other contributors:
\itemize{
  \item Ernest Fokoue \email{epfeqa@rit.edu} [contributor]
  \item Golibe Ezeechesi \email{golibe.ezeechesi@gmail.com} [contributor]
  \item Fatimo Adebanjo \email{adebanjofatimo2000@gmail.com} [contributor]
}

}
\keyword{internal}
